Successfully accessing Twiter API and getting home timeline
On clicking the button on the login screen, a web browser app is
launched with a Twitter authorisation page in which the user can
enter their username and password in order to authorise our app
to make access the Twitter API on their behalf (e.g. reading their
home timeline, i.e. own posts and posts of followed people).

From this point on, we apparently have an access token which is
sent with every request, and every request is specific to this user
that initially authorised our app.

We can then make all the API calls as defined on
https://dev.twitter.com/rest/reference

Bad thing: the Codepath android-oauth-handler
(https://github.com/codepath/android-oauth-handler) which is used
by this app skeleton uses Android-Async-HTTP which cannot be
combined with Stetho for monitoring the HTTP API requests and
responses.

Maybe it would be better to only use ScribeJava for the OAuth stuff
and then use Retrofit with OkHttp for the HTTP requests, which can
be combined with Stetho.