Save current user in SharedPreferences in TimelineActivity
The currently authenticated user is queried on activity creation
and saved in SharedPreferences. If the app is opened offline
(offline mode), then the current user is read from the SharedPrefs
rather than from the server.

Also. pass the current user to the ComposeActivity so that it
does not need to be fetched from the server again there.