Pass offline mode to ComposeActivity
If the ComposeActivity is launched in offline mode, the input
widgets (EditText and Button) are disabled, and the offline
indicator bar is displayed.

Also: whenever the first page of tweets is loaded, the current
time is saved in the SharedPreferences, and this time is
displayed as "last update" in the offline indicator in
TimelineActivity.