Introduce OfflineModeManager; improve LoginManager
The LoginManager is better integrated and has a method for
determining the current user, and a callback for success and failure
that must be passed to this method.

The LoginManager and OfflineModeManager can be called from anywhere
in the application for determining the current user and whether
the app is in online or offline mode. However, it's necessary that
both managers are properly initialised at some point ("determineX"
methods).