Launch ProfileActivity on any profile image click in RecyclerView
The click on the profile image is detected in the adapter for the
RecyclerView items. From there it is communicated via a listener
callback (the adapter has a setOnProfileImageClickListener method
used by the TimelineFragment) to the TimelineFragment. In this
callback, the fragment communicates the clicked user via another
callback (the TimelineFragment has a TimelineFragmentListener
implemented by the TimelineActivity) to the TimelineActivity.
Finally, the TimelineActivity launches the ProfileActivity by
passing the User object whose profile image has been clicked.