Delete only tweets belonging to a certain timeline on loading page 1
Now, if the same tweet is fetched for a different timeline (e.g.
home timeline instead of user timeline), then the "type" field of
this tweet will be overriden in the database, i.e. the full
timelines are not maintained in the database. This problem could
be solved by  making the primary key a composite of the tweet ID
and the type.

The reason for keeping similar tweets belonging to a different
timeline in the database is that the different timelines can be
restored in offline mode.